Migrate hand-capture legacy titles to shared renderLegacyTitle util
Replace the per-sketch nested text/overlay { show, title, subtitle } blocks
with flat title/subtitle params in each hand-capture sketch, rendered by a
new reusable @/p5/utils/title/renderLegacyTitle.js (extracted verbatim from
HandCaptureScene.drawTitle, now removed from _shared.js). Also drop dead
commented references to the pre-content-item title system.
