+++ src/sketches/p5/sketches/hand-capture/hand-tracking-v0-move/index.js
@@ -4,6 +4,7 @@ import sketch from "@/p5/utils/sketch.js";
 import {
   HandCaptureScene
 } from "../_shared.js";
+import renderLegacyTitle from "@/p5/utils/title/renderLegacyTitle.js";
 
 const scene = new HandCaptureScene( {
   layers: {
@@ -21,7 +22,6 @@ sketch.draw( () => {
   const physics = options.sketch?.physics ?? {};
   const visuals = options.sketch?.visuals ?? {};
   const hands = options.sketch?.hands ?? {};
-  const text = options.sketch?.text ?? {};
   const background = options.sketch?.backgroundColor ?? options.colors?.background ?? [
     0
   ];
@@ -55,10 +55,9 @@ sketch.draw( () => {
   } );
   scene.compose();
 
-  scene.drawTitle( {
-    title: text.title ?? "move",
-    subtitle: text.subtitle ?? "hand tracking v0",
-    show: text.show ?? true,
+  renderLegacyTitle( {
+    title: options.sketch?.title ?? "",
+    subtitle: options.sketch?.subtitle ?? "",
     color: background
   } );
 } );
