+++ src/sketches/p5/sketches/hand-capture/hand-tracking-v0-move/options.ts
@@ -65,12 +65,9 @@ export const formValues = {
     trail: 255
   },
 
-  // Overlay text
-  text: {
-    show: true,
-    title: "move",
-    subtitle: "hand tracking v0"
-  }
+  // Legacy centred splash (rendered by @/p5/utils/title/renderLegacyTitle.js)
+  title: "move",
+  subtitle: "hand tracking v0"
 };
 
 // UI configuration only
@@ -182,22 +179,13 @@ export const formConfiguration: Record<string, any> = {
     }
   },
 
-  text: {
-    component: "nested-object",
-    label: "Overlay text",
-    fields: {
-      show: {
-        label: "Show text",
-        component: "checkbox"
-      },
-      title: {
-        label: "Title",
-        component: "text"
-      },
-      subtitle: {
-        label: "Subtitle",
-        component: "text"
-      }
-    }
+  title: {
+    label: "Title",
+    component: "text"
+  },
+
+  subtitle: {
+    label: "Subtitle",
+    component: "text"
   }
 };
