+++ src/sketches/p5/sketches/hand-capture/hand-tracking-v1-attract/options.ts
@@ -63,11 +63,9 @@ export const formValues = {
     trail: 255
   },
 
-  text: {
-    show: true,
-    title: "attract",
-    subtitle: "hand tracking v1"
-  }
+  // Legacy centred splash (rendered by @/p5/utils/title/renderLegacyTitle.js)
+  title: "attract",
+  subtitle: "hand tracking v1"
 };
 
 export const formConfiguration: Record<string, any> = {
@@ -192,22 +190,13 @@ export const formConfiguration: Record<string, any> = {
     }
   },
 
-  text: {
-    component: "nested-object",
-    label: "Overlay text",
-    fields: {
-      show: {
-        label: "Show text",
-        component: "checkbox"
-      },
-      title: {
-        label: "Title",
-        component: "text"
-      },
-      subtitle: {
-        label: "Subtitle",
-        component: "text"
-      }
-    }
+  title: {
+    label: "Title",
+    component: "text"
+  },
+
+  subtitle: {
+    label: "Subtitle",
+    component: "text"
   }
 };
