+++ src/sketches/p5/sketches/hand-capture/hand-tracking-v2-repulse/index.js
@@ -4,6 +4,7 @@ import sketch from "@/p5/utils/sketch.js";
 import {
   HandCaptureScene
 } from "../_shared.js";
+import renderLegacyTitle from "@/p5/utils/title/renderLegacyTitle.js";
 
 const scene = new HandCaptureScene( {
   layers: {
@@ -22,7 +23,6 @@ sketch.draw( () => {
   const visuals = options.sketch?.visuals ?? {};
   const hands = options.sketch?.hands ?? {};
   const repulse = options.sketch?.repulse ?? {};
-  const text = options.sketch?.text ?? {};
   const background = options.sketch?.backgroundColor ?? options.colors?.background ?? [
     0
   ];
@@ -57,10 +57,9 @@ sketch.draw( () => {
   } );
   scene.compose();
 
-  scene.drawTitle( {
-    title: text.title ?? "repulse",
-    subtitle: text.subtitle ?? "hand tracking v2",
-    show: text.show ?? true,
+  renderLegacyTitle( {
+    title: options.sketch?.title ?? "",
+    subtitle: options.sketch?.subtitle ?? "",
     color: background
   } );
 } );
