+++ src/sketches/p5/sketches/hand-capture/hand-tracking-v6-slice/options.ts
@@ -75,11 +75,9 @@ export const formValues = {
     show: true
   },
 
-  // Overlay text
-  text: {
-    show: true,
-    subtitle: "slice · hand tracking v6"
-  }
+  // Legacy centred splash (rendered by @/p5/utils/title/renderLegacyTitle.js)
+  title: "",
+  subtitle: "slice · hand tracking v6"
 };
 
 // UI configuration only
@@ -244,18 +242,13 @@ export const formConfiguration: Record<string, any> = {
     }
   },
 
-  text: {
-    component: "nested-object",
-    label: "Overlay text",
-    fields: {
-      show: {
-        label: "Show text",
-        component: "checkbox"
-      },
-      subtitle: {
-        label: "Subtitle",
-        component: "text"
-      }
-    }
+  title: {
+    label: "Title",
+    component: "text"
+  },
+
+  subtitle: {
+    label: "Subtitle",
+    component: "text"
   }
 };
