+++ src/sketches/p5/sketches/hand-capture/hand-tracking-v7-echo/options.ts
@@ -52,12 +52,9 @@ export const formValues = {
     ghostAlpha: 0.55
   },
 
-  // Overlay text — centered title (like the move / attract / restore sketches)
-  text: {
-    show: true,
-    title: "echo",
-    subtitle: "hand tracking v7"
-  }
+  // Legacy centred splash (rendered by @/p5/utils/title/renderLegacyTitle.js)
+  title: "echo",
+  subtitle: "hand tracking v7"
 };
 
 // UI configuration only
@@ -133,22 +130,13 @@ export const formConfiguration: Record<string, any> = {
     }
   },
 
-  text: {
-    component: "nested-object",
-    label: "Overlay text",
-    fields: {
-      show: {
-        label: "Show text",
-        component: "checkbox"
-      },
-      title: {
-        label: "Title",
-        component: "text"
-      },
-      subtitle: {
-        label: "Subtitle",
-        component: "text"
-      }
-    }
+  title: {
+    label: "Title",
+    component: "text"
+  },
+
+  subtitle: {
+    label: "Subtitle",
+    component: "text"
   }
 };
