+++ src/sketches/p5/sketches/photo/_photo-letter-gravity/index.js
@@ -381,7 +381,6 @@ sketch.draw( (
   // if (animation.progression < 0.2) {
   // 	string.write(
   // 		defaultTitle,
-  // 		// options.texts.title || defaultTitle,
   // 		p.width/2,
   // 		p.height/2,
   // 		{
