+++ src/sketches/p5/sketches/text/text-decomposition/index.js
@@ -178,6 +178,4 @@ sketch.draw( () => {
   //     popPush: true,
   //   }
   // );
-
-  // renderTitle( options.sketch?.text ?? options.name );
 } );
